Split tool buttons out of SLDTree class into SLDTreeTools
